+++ src/main/java/com/github/idea/ginkgo/GinkgoConsoleProperties.java
@@ -1,6 +1,5 @@
 package com.github.idea.ginkgo;
 
-import com.goide.execution.testing.frameworks.gotest.GotestEventsConverter;
 import com.intellij.execution.Executor;
 import com.intellij.execution.testframework.TestConsoleProperties;
 import com.intellij.execution.testframework.sm.SMCustomMessagesParsing;